+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 # Change Log
 
 ## [Unreleased]
+
+- Add pause screen api
+
 ## [1.4.1] 2025-05-04
 
 ### Fixed